Added: admin get bookings for period Separated the bookings operations logic between admin and client

+++ src/main/java/vk/haveplace/services/objects/requests/ShiftRequest.java
@@ -7,7 +7,7 @@ import org.aspectj.weaver.ast.Not;
 @Data
 public class ShiftRequest {
     @NotNull
-    private int adminId;
+    private long adminId;
     @NotNull
     private DateAndTimesRequest shift;
 }
